+++ packages/datasource-sql/test/introspection/introspector.integration.test.ts
@@ -1,4 +1,3 @@
-/* eslint-disable max-len */
 import { DataTypes, Sequelize, literal } from 'sequelize';
 
 import Introspector from '../../src/introspection/introspector';