+++ packages/plugin-flattener/src/flatten-json-column/index.ts
@@ -81,7 +81,6 @@ export default function flattenJsonColumn(
 
   if (!Object.keys(options.columnType).length) {
     throw new Error(
-      // eslint-disable-next-line max-len
       'options.columnType must be defined as json object representing a subset of the shape of the data in the json column.',
     );
   }