+++ packages/_example/CHANGELOG.md
@@ -1,3 +1,25 @@
+## example [1.8.5](https://github.com/ForestAdmin/agent-nodejs/compare/example@1.8.4...example@1.8.5) (2024-04-08)
+
+
+
+
+
+### Dependencies
+
+* **@forestadmin/agent:** upgraded to 1.38.10
+* **@forestadmin/datasource-dummy:** upgraded to 1.1.11
+
+## example [1.8.4](https://github.com/ForestAdmin/agent-nodejs/compare/example@1.8.3...example@1.8.4) (2024-04-05)
+
+
+
+
+
+### Dependencies
+
+* **@forestadmin/agent:** upgraded to 1.38.9
+* **@forestadmin/datasource-dummy:** upgraded to 1.1.10
+
 ## example [1.8.3](https://github.com/ForestAdmin/agent-nodejs/compare/example@1.8.2...example@1.8.3) (2024-04-04)